Cleaned up some warnings; refactored, renamed some tests; edited travis-ci file.
