+++ validator-core/pom.xml
@@ -95,7 +95,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.apache.commons</groupId>
-			<artifactId>commons-lang3</artifactId>
+			<artifactId>commons-text</artifactId>
 		</dependency>
 	</dependencies>
